15829: fix for large

+++ problem/15829/main.cpp
@@ -4,8 +4,8 @@
 
 using namespace std;
 
-const int r = 31;
-const int M = 1234567891;
+const long long r = 31;
+const long long M = 1234567891;
 int L;
 string S;
 
@@ -13,7 +13,7 @@ int main() {
     ios_base::sync_with_stdio(false);
     cin.tie(NULL);
     cin >> L >> S;
-    int hash_val = 0;
+    long long hash_val = 0;
     for (auto it = S.rbegin(); it != S.rend(); it++) {
         hash_val = hash_val * r + *it - 'a' + 1;
         hash_val %= M;
